Consolidate unit tests
